# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: POS + Inventory + Delivery Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: getProducts
      tags: [products]
      summary: List products
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductInput"
      responses:
        "201":
          description: Created product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    put:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductInput"
      responses:
        "200":
          description: Updated product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Invoices ─────────────────────────────────────────────────────────────────
  /invoices:
    get:
      operationId: getInvoices
      tags: [invoices]
      summary: List invoices
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: dateFrom
          in: query
          schema:
            type: string
        - name: dateTo
          in: query
          schema:
            type: string
        - name: deliveryId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of invoices
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invoice"
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create an invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceInput"
      responses:
        "201":
          description: Created invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
  /invoices/{id}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get invoice with items
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice with items
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceWithItems"
    put:
      operationId: updateInvoice
      tags: [invoices]
      summary: Update an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceInput"
      responses:
        "200":
          description: Updated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /invoices/{id}/duplicate:
    post:
      operationId: duplicateInvoice
      tags: [invoices]
      summary: Duplicate an invoice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Duplicated invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoiceWithItems"
  /invoices/last-price:
    get:
      operationId: getLastSalePrice
      tags: [invoices]
      summary: Get last sale price for a product+customer pair
      parameters:
        - name: customerName
          in: query
          required: true
          schema:
            type: string
        - name: productId
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Last sale info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LastSaleInfo"

  # ── Customer Purchase History ──────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        categoryId:
          type: integer
          nullable: true
        categoryName:
          type: string
          nullable: true
        basePrice:
          type: number
        trackStock:
          type: boolean
        stockQty:
          type: integer
        createdDate:
          type: string
      required: [id, name, basePrice, trackStock, stockQty, createdDate]
    CreateProductInput:
      type: object
      properties:
        name:
          type: string
        categoryId:
          type: integer
          nullable: true
        basePrice:
          type: number
        trackStock:
          type: boolean
        stockQty:
          type: integer
        createdDate:
          type: string
      required: [name, basePrice, trackStock, stockQty]
    InvoiceItem:
      type: object
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        qty:
          type: number
        price:
          type: number
        subtotal:
          type: number
        isDamaged:
          type: boolean
      required: [id, invoiceId, productId, productName, qty, price, subtotal, isDamaged]
    InvoiceItemInput:
      type: object
      properties:
        productId:
          type: integer
        qty:
          type: number
        price:
          type: number
      required: [productId, qty, price]
    Invoice:
      type: object
      properties:
        id:
          type: integer
        invoiceNo:
          type: string
        customerName:
          type: string
        date:
          type: string
        total:
          type: number
        deliveryId:
          type: integer
          nullable: true
        deliveryNo:
          type: string
          nullable: true
        note:
          type: string
          nullable: true
      required: [id, invoiceNo, customerName, date, total]
    InvoiceWithItems:
      allOf:
        - $ref: "#/components/schemas/Invoice"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/InvoiceItem"
          required: [items]
    CreateInvoiceInput:
      type: object
      properties:
        customerName:
          type: string
        date:
          type: string
        deliveryId:
          type: integer
          nullable: true
        note:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItemInput"
      required: [customerName, date, items]
    LastSaleInfo:
      type: object
      properties:
        price:
          type: number
        date:
          type: string
          nullable: true
        hasHistory:
          type: boolean
      required: [price, hasHistory]
...