# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * POS + Inventory + Delivery Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  categoryId?: number | null;
  categoryName?: string | null;
  basePrice: number;
  trackStock: boolean;
  stockQty: number;
  createdDate: string;
}